requirements.txt
fastapi
uvicorn
jinja2
torch
torchvision
Pillow
numpy
efficientnet_pytorch
python-multipart